changed storage of Players in MenuModel and locations of methods.
Instead of two separate Player objects, they are now stored in an array
and can be accessed based on their playerNum. changeBoardSize has been
moved to the Model since it makes more sense to be in the model. The
changePlayerColor method is now also in the model for the same reason.